Refactor common elements from protocol into superclasses
