+++ sail_on_client/protocol/condda_test.py
@@ -1,25 +1,22 @@
 """Test for CONDDA."""
 
 import logging
-import os
-import pickle as pkl
 from itertools import count
-from typing import Union, Tuple, Dict, Any, List
-import ubelt as ub
+from typing import Union, Dict, List
 
 from sail_on_client.protocol.condda_dataclasses import AlgorithmAttributes, InitializeParams
+from sail_on_client.protocol.visual_test import VisualTest
 from sail_on_client.protocol.condda_round import CONDDARound
 from sail_on_client.protocol.parinterface import ParInterface
 from sail_on_client.protocol.localinterface import LocalInterface
 from sail_on_client.utils.utils import safe_remove
-from sail_on_client.utils.decorators import skip_stage
 from sail_on_client.errors import RoundError
 
 
 log = logging.getLogger(__name__)
 
 
-class CONDDATest:
+class CONDDATest(VisualTest):
     """Class representing CONDDA Test."""
 
     def __init__(
@@ -33,9 +30,8 @@ class CONDDATest:
             skip_stages: List[str],
             use_consolidated_features: bool,
             use_saved_features: bool) -> None:
-
         """
-        Constructor for test for CONDDA.
+        Construct test for CONDDA.
 
         Args:
             algorithm_attributes: An instance of algorithm_attributes
@@ -51,89 +47,15 @@ class CONDDATest:
         Returns:
             None
         """
-        self.algorithm_attributes = algorithm_attributes
-        self.data_root = data_root
-        self.domain = domain
-        self.harness = harness
-        self.save_dir = save_dir
-        self.session_id = session_id
-        self.skip_stages = skip_stages
-        self.use_consolidated_features = use_consolidated_features
-        self.use_saved_features = use_saved_features
-
-    def _restore_features(
-            self,
-            test_id: str) -> Tuple[Dict, Dict]:
-        """
-        Private function to _restore_features.
-
-        Args:
-           test_id: An identifier for the test
-
-        Returns:
-            Tuple of dictionary with features and logits obtained from the feature extractor
-        """
-        features_dict: Dict = {}
-        logit_dict: Dict = {}
-        algorithm_name = self.algorithm_attributes.name
-        if self.use_saved_features:
-            if os.path.isdir(self.save_dir):
-                if self.use_consolidated_features:
-                    feature_fname = f"{algorithm_name}_features.pkl"
-                else:
-                    feature_fname = f"{test_id}_{algorithm_name}_features.pkl"
-                feature_path = os.path.join(self.save_dir, feature_fname)
-                test_features = pkl.load(open(feature_path, "rb"))
-            else:
-                test_features = pkl.load(open(self.save_dir, "rb"))
-            features_dict = test_features["features_dict"]
-            logit_dict = test_features["logit_dict"]
-        return features_dict, logit_dict
-
-    def _aggregate_features_across_round(
-            self,
-            round_instance: CONDDARound,
-            feature_dict: Dict,
-            logit_dict: Dict) -> Tuple[Dict, Dict]:
-        """
-        Aggregate features across multiple rounds.
-
-        Args:
-            round_instance: Instance of ond round
-            feature_dict: Aggregated features until this function was called
-            logit_dict: Aggregated logit until this function was called
-
-        Return:
-            Tuple of features and logits with features and logits from the round
-        """
-        feature_dict.update(getattr(round_instance, "rfeature_dict", {}))
-        logit_dict.update(getattr(round_instance, "rlogit_dict", {}))
-        return feature_dict, logit_dict
-
-    @skip_stage("SaveFeatures")
-    def _save_features(self,
-                       test_id: str,
-                       feature_dict: Dict,
-                       logit_dict: Dict) -> None:
-        """
-        Save features for a test.
-
-        Args:
-            test_id: An identifier for the test
-            feature_dict: Features for the test
-            logit_dict: Logit for the test
-
-        Return:
-            None
-        """
-        ub.ensuredir(self.save_dir)
-        algorithm_name = self.algorithm_attributes.name
-        feature_path = os.path.join(
-            self.save_dir, f"{test_id}_{algorithm_name}_features.pkl"
-        )
-        log.info(f"Saving features in {feature_path}")
-        with open(feature_path, "wb") as f:
-            pkl.dump({"features_dict": feature_dict, "logit_dict": logit_dict}, f)
+        super().__init__(algorithm_attributes,
+                         data_root,
+                         domain,
+                         harness,
+                         save_dir,
+                         session_id,
+                         skip_stages,
+                         use_consolidated_features,
+                         use_saved_features)
 
     def __call__(self, test_id: str) -> None:
         """
