+++ tests/test_ond_dataclasses.py
@@ -4,7 +4,13 @@ import pytest
 import uuid
 from sail_on_client import __version__
 from sail_on_client.mock import MockDetector
-from sail_on_client.protocol.ond_dataclasses import AlgorithmAttributes, InitializeParams, FeatureExtractionParams, WorldChangeDetectionParams, NoveltyClassificationParams, NoveltyAdaptationParams, NoveltyCharacterizationParams
+from sail_on_client.protocol.ond_dataclasses import (
+        AlgorithmAttributes,
+        InitializeParams,
+        NoveltyClassificationParams,
+        NoveltyAdaptationParams,
+        NoveltyCharacterizationParams
+)
 
 
 @pytest.fixture(scope="function")
@@ -31,27 +37,6 @@ def initialize_params():
         "session_id": str(uuid.uuid4()),
         "test_id": "OND.10.90001.2100554",
         "feedback_instance": None
-    }
-
-
-@pytest.fixture(scope="function")
-def feature_extraction_params():
-    """Fixture for providing parameters for initialize FeatureExtractionParams."""
-    return {
-        "dataset": "",
-        "data_root": "",
-        "redlight_image": "test.png",
-        "round_id": 0
-    }
-
-
-@pytest.fixture(scope="function")
-def world_change_detection_params():
-    """Fixture for providing parameters for initialize WorldChangeDetectionParams."""
-    return {
-        "features_dict": {},
-        "logit_dict": {},
-        "round_id": 0
     }
 
 
@@ -172,69 +157,6 @@ def test_initialize_params_get_toolset(initialize_params):
                                          "session_id": initialize_params["session_id"],
                                          "test_id": "OND.10.90001.2100554",
                                          "test_type": ""}
-
-
-def test_feature_extraction_initialize(feature_extraction_params):
-    """
-    Test FeatureExtractionParams __init__.
-
-    Args:
-        feature_extraction_params: Dictionary with parameters to initialize FeatureExtractionParams
-
-    Returns:
-        None
-    """
-    FeatureExtractionParams(**feature_extraction_params)
-
-
-def test_feature_extraction_params_get_toolset(feature_extraction_params):
-    """
-    Test FeatureExtractionParams get_toolset.
-
-    Args:
-        feature_extraction_params: Dictionary with parameters to initialize FeatureExtractionParams
-
-    Returns:
-        None
-    """
-    fe_params = FeatureExtractionParams(**feature_extraction_params)
-    assert fe_params.get_toolset() == {
-            "dataset": "",
-            "dataset_root": "",
-            "redlight_image": "test.png",
-            "round_id": 0
-            }
-
-
-def test_world_change_detection_initialize(world_change_detection_params):
-    """
-    Test WorldChangeDetectionParams __init__.
-
-    Args:
-        world_change_detection_params: Dictionary with parameters to initialize WorldChangeDetectionParams
-
-    Returns:
-        None
-    """
-    WorldChangeDetectionParams(**world_change_detection_params)
-
-
-def test_world_change_detection_get_toolset(world_change_detection_params):
-    """
-    Test WorldChangeDetectionParams get_toolset.
-
-    Args:
-        world_change_detection_params: Dictionary with parameters to initialize WorldChangeDetectionParams
-
-    Returns:
-        None
-    """
-    wc_params = WorldChangeDetectionParams(**world_change_detection_params)
-    assert wc_params.get_toolset() == {
-            "features_dict": {},
-            "logit_dict": {},
-            "round_id": 0
-           }
 
 
 def test_novelty_classification_initialize(novelty_classification_params):
